requirements.txt
binwalk==2.1.0
decorator==4.4.2
networkx==2.5.1
pkg-resources==0.0.0
pygraphviz==1.6
python-magic==0.4.22
